Threading of the detect module has been disabled
The detect module threading is disabled since it can only be run once before
requiring the python interpreter to be restarted (restarting the GUI alone is
not enough). The fact that the GUI can run the thread once bodes well, but
the reason for failing to subsequently run detection is unsure. It returns
a main loop error, which suggests that the error could be due to the
thread.start() call not being main in the main tkinter GUI loop. However
the test module deepspace.py returns no such error, despite using the same
threading method as the threaded detect module. Future work will hopefully
determine test starting the detection thread in the main loop in hopes
that it will allow for continual calling of the detect module and
iterative display of detection results.